+++ BetterCrafting/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.3.0
-Released ???.
+Released September 20th, 2022.
 
 ### New Features